Drop `yum_requirements.txt`
To build `openspecfunc`, we need a Fortran compiler. As conda-forge had
not previously included a Fortran compiler in the
`condaforge/linux-anvil` image, the `yum_requirements.txt` file was
needed to install the Fortran compiler. However now the
`condaforge/linux-anvil` image includes a Fortran compiler. So this is
no longer needed. Hence drop the `yum_requirements.txt` file.